gprofng: fix build with -mx32
gprofng/ChangeLog
2022-05-27  Vladimir Mezentsev  <vladimir.mezentsev@oracle.com>

	PR gprofng/28983
	PR gprofng/29143
	* src/Experiment.cc (write_header): Fix argument for ctime.
	Fix -Wformat= warnings.
	* src/Dbe.cc: Likewise.
	* src/DwarfLib.h: Fix [-Wsign-compare] warnings.
	* src/Experiment.h: Likewise.
	* src/ipc.cc: Fix -Wformat= warnings.

+++ gprofng/src/ipc.cc
@@ -1281,7 +1281,7 @@ ipc_doWork (void *arg)
     {
       int arg1 = readInt (req);
       uint64_t arg2 = readLong (req);
-      ipc_log ("  args = %d, %ld\n", arg1, arg2);
+      ipc_log ("  args = %d, %lld\n", arg1, (long long) arg2);
       dbeSetSelObjV2 (arg1, arg2);
       writeString (NULL, req);
     }
